ranging send result to ARM (STEP 2 finished)

+++ branches/sugar/sensorfly4_1/SF_Network/sf_network.h
@@ -36,7 +36,7 @@ typedef struct
   uint8_t dest; // with 6 characters MAC Address
   uint8_t src;	// with 6 characters MAC Address
   uint8_t length;
-  int16_t data[10];
+  double data[10];
 } Packet __attribute__((packed));
